+++ webrtc/common_audio/blocker_unittest.cc
@@ -18,12 +18,12 @@ namespace {
 class PlusThreeBlockerCallback : public webrtc::BlockerCallback {
  public:
   void ProcessBlock(const float* const* input,
-                    int num_frames,
+                    size_t num_frames,
                     int num_input_channels,
                     int num_output_channels,
                     float* const* output) override {
     for (int i = 0; i < num_output_channels; ++i) {
-      for (int j = 0; j < num_frames; ++j) {
+      for (size_t j = 0; j < num_frames; ++j) {
         output[i][j] = input[i][j] + 3;
       }
     }
@@ -34,12 +34,12 @@ class PlusThreeBlockerCallback : public webrtc::BlockerCallback {
 class CopyBlockerCallback : public webrtc::BlockerCallback {
  public:
   void ProcessBlock(const float* const* input,
-                    int num_frames,
+                    size_t num_frames,
                     int num_input_channels,
                     int num_output_channels,
                     float* const* output) override {
     for (int i = 0; i < num_output_channels; ++i) {
-      for (int j = 0; j < num_frames; ++j) {
+      for (size_t j = 0; j < num_frames; ++j) {
         output[i][j] = input[i][j];
       }
     }
